# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WritingsReader API",
    "version": "0.4.0"
  },
  "paths": {
    "/api/writings": {
      "post": {
        "tags": [
          "Writings"
        ],
        "summary": "Create Writing",
        "operationId": "create_writing_api_writings_post",
        "parameters": [
          {
            "name": "x-device-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWritingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingDetailResponse"
                }
              }
            }
          },
          "200": {
            "description": "An identical publish request already succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingDetailResponse"
                }
              }
            }
          },
          "409": {
            "description": "The client ID is already used by different content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid X-Device-ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Writing persistence failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Writings"
        ],
        "summary": "List Writings",
        "operationId": "list_writings_api_writings_get",
        "parameters": [
          {
            "name": "x-device-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WritingPreviewResponse"
                  },
                  "title": "Response List Writings Api Writings Get"
                }
              }
            }
          },
          "400": {
            "description": "Invalid X-Device-ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Writing persistence failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/writings/{writing_id}": {
      "get": {
        "tags": [
          "Writings"
        ],
        "summary": "Get Writing",
        "operationId": "get_writing_api_writings__writing_id__get",
        "parameters": [
          {
            "name": "writing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Writing Id"
            }
          },
          {
            "name": "x-device-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingDetailResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "400": {
            "description": "Invalid X-Device-ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Writing persistence failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/censor": {
      "post": {
        "tags": [
          "Writings"
        ],
        "summary": "Censor Text",
        "operationId": "censor_text_api_censor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CensorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CensorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CensorRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 50000,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CensorRequest"
      },
      "CensorResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "was_censored": {
            "type": "boolean",
            "title": "Was Censored"
          }
        },
        "type": "object",
        "required": [
          "text",
          "was_censored"
        ],
        "title": "CensorResponse"
      },
      "CreateWritingRequest": {
        "properties": {
          "client_writing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Writing Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "full_text": {
            "type": "string",
            "maxLength": 50000,
            "title": "Full Text"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptSnapshot"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_writing_id",
          "full_text",
          "prompt"
        ],
        "title": "CreateWritingRequest"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PromptSnapshot": {
        "properties": {
          "verb": {
            "type": "string",
            "title": "Verb"
          },
          "full_text": {
            "type": "string",
            "title": "Full Text"
          },
          "emotions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Emotions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verb",
          "full_text",
          "emotions"
        ],
        "title": "PromptSnapshot"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WritingDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "client_writing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Writing Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "full_text": {
            "type": "string",
            "title": "Full Text"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptSnapshot"
          },
          "status": {
            "type": "string",
            "const": "published",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_writing_id",
          "title",
          "full_text",
          "author_id",
          "prompt",
          "status",
          "created_at",
          "published_at"
        ],
        "title": "WritingDetailResponse"
      },
      "WritingPreviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "client_writing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Writing Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "preview_text": {
            "type": "string",
            "title": "Preview Text"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptSnapshot"
          },
          "status": {
            "type": "string",
            "const": "published",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_writing_id",
          "title",
          "preview_text",
          "author_id",
          "prompt",
          "status",
          "created_at",
          "published_at"
        ],
        "title": "WritingPreviewResponse"
      }
    }
  }
}